+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.1.0
+
+- Select multiple notes by long pressing
+- After entering select mode, add notes to the selection by tapping them
+- Select or unselect all notes in the select menu
+- Favorite/Unfavorite multiple notes at once
+- Pin/Unpin multiple notes at once
+- Add and remove tags to multiple notes at once
+- Move to Trash and delete multiple notes at once
+
 ## 0.0.9
 
 - Dark Theme